requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
Flask==1.1.1
Django==2.0.3
requests
boto3
PyMySQL==0.9.3
scikit-learn==0.22.1
tensorflow==2.0.0
keras==2.3.1
scipy==1.4.1
pandas==1.0.1

